% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{add_coords}
\alias{add_coords}
\title{Create Coordinates for GSLIB-style Grid file.}
\usage{
add_coords(data, grid_def)
}
\arguments{
\item{data}{Data frame with grid arrangements but no coordiates. Imported
from a GeoEase system file with function like \code{read_gslib}.}

\item{grid_def}{Standard grid definition, which is a named numeric vectoor
with elements (in order): n_x, n_y, n_z, min_x, min_y, min_z, dim_x,
dim_y, dim_z, realz.}
}
\value{
Data frame same as input but with fields: r, x, y, z added.
}
\description{
Create Coordinates for GSLIB-style Grid file.
}
