% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addcoordR.R
\name{addcoordR}
\alias{addcoordR}
\title{Run the CCG \code{addcoord} Program to Create a Variogram Map.}
\usage{
addcoordR(path, realization = 1, xdef = c(50, 2.5, 5), ydef = c(50,
  2.5, 5), zdef = c(1, 0.5, 1), dec = c(3, 3, 3), debug = FALSE)
}
\arguments{
\item{path}{GeoEase format system file.}

\item{realization}{Scalar numeric. Number of the realization to affect.}

\item{xdef, ydef, zdef}{Numeric vectors of length 3 defining the number of
nodes, grid origin coordinate, and node spacing for x, y, and z axes.}

\item{dec}{Rounding of coordinates for x, y, and z.}

\item{debug}{Scalar boolean. If \code{TRUE} don't delete temporary system
files. Default is \code{FALSE}.}
}
\value{
Silently return \code{path}. Side effect is creation of system file
called 'addcoord-out.dat' which is the input \code{path} file with the
addition of coordinates.
}
\description{
\code{addcoordR} creates variogram maps for analysis of spatial continuity.
}
