% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{build_grid_def}
\alias{build_grid_def}
\title{Build a Grid Defintion from Stored Parameters}
\usage{
build_grid_def(data, domain)
}
\arguments{
\item{data}{Data frame of stored parameters with specific numeric columns:
domain, nx, ny, nz, min_x, min_y, min_z, dim_x, dim_y, dim_z}

\item{domain}{Scalar numeric or character (matched to domain in \code{data})
domain, category, or zone code.}
}
\value{
A named numeric vector defining a grid.
}
\description{
Build a Grid Defintion from Stored Parameters
}
