% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{build_vario_par_list}
\alias{build_vario_par_list}
\title{Build a Variogram Calculation Parameter List from Stored Parameters}
\usage{
build_vario_par_list(data, domain, var, ndims)
}
\arguments{
\item{data}{Data frame of stored parameters (see example for strict
structure of the data frame).}

\item{domain}{Scalar numeric or character (matched to type in \code{data})
domain, category, or zone code.}

\item{var}{Scalar character variable name for variogram.}

\item{ndims}{Scalar numeric 1--3: number of variogram dimensions (axes) with
1 being major; 2 semi-major; and 3 minor.}
}
\value{
A list of axis variogram parameters suitable for input into variogram
calculation function \link{varcalcR}.
}
\description{
Build a Variogram Calculation Parameter List from Stored Parameters
}
\examples{
vario_calc <- tibble::tribble(
~domain, ~vars, ~axis, ~azm, ~azmtol, ~bndhorz, ~dip, ~diptol, ~bndvert,
~tilt, ~nlags, ~dlag, ~lagtol,
38, "NS_thk", 1, 90, 45, 1.0e21, 0, 90, 1.0e21, 0, 15, 10, 0.5,
38, "NS_thk", 2, 180, 45, 1.0e21, 0, 90, 1.0e21, 0, 10, 10, 0.5,
38, "NS_accum", 1, 90, 45, 1.0e21, 0, 90, 1.0e21, 0, 15, 10, 0.5,
38, "NS_accum", 2, 180, 45, 1.0e21, 0, 90, 1.0e21, 0, 10, 10, 0.5,
38, "NS_thkaccum", 1, 90, 45, 1.0e21, 0, 90, 1.0e21, 0, 15, 10, 0.5,
38, "NS_thkaccum", 2, 180, 45, 1.0e21, 0, 90, 1.0e21, 0, 10, 10, 0.5
)
build_vario_par_list(vario_calc, 38, "NS_accum", 2)
}
