% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{create_grid}
\alias{create_grid}
\title{Populate a Data Fram Grid Representation Using a Grid Definition}
\usage{
create_grid(grid_def)
}
\arguments{
\item{grid_def}{Standard grid definition, which is a named numeric vecotor
with elements (in order): n_x, n_y, n_z, min_x, min_y, min_z, dim_x,
dim_y, dim_z, realz.}
}
\value{
A data frame with grid coordinates.
}
\description{
Populate a Data Fram Grid Representation Using a Grid Definition
}
