% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-write.R
\name{create_gslib_griddef}
\alias{create_gslib_griddef}
\title{Create a GSLIB Grid Definition from a Grid Data Frame.}
\usage{
create_gslib_griddef(data, dims = c(1, 1), xyz = c("x", "y"),
  realz = 1)
}
\arguments{
\item{data}{Data frame representing grid.}

\item{dims}{Numeric vector of xyz grid node spacing. Length 2 for 2D.}

\item{xyz}{Character vector of xyz coordinate column names. Length 2 for 2D.}

\item{realz}{Number of realizations.}
}
\value{
Named numeric vector of grid definition.
}
\description{
Create a GSLIB Grid Definition from a Grid Data Frame.
}
