% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{get_column_indices}
\alias{get_column_indices}
\title{Get Vector of Specific Column Indices of from a Data Frame.}
\usage{
get_column_indices(data, vars)
}
\arguments{
\item{data}{Data frame}

\item{vars}{Character vector of column names in \code{data}.}
}
\value{
Numeric vector of column indices.
}
\description{
Get Vector of Specific Column Indices of from a Data Frame.
}
