% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_export.R
\name{gslib2tibble}
\alias{gslib2tibble}
\title{Import GeoEase format file to a tibble.}
\usage{
gslib2tibble(file)
}
\arguments{
\item{file}{character, name and path of GeEase format data file.}
}
\value{
a tibble
}
\description{
Import GeoEase format file to a tibble.
}
