% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histpltsimR.R
\name{histpltsimR}
\alias{histpltsimR}
\title{Generate Histogram Data for Simulations}
\usage{
histpltsimR(sims, simvar, samples, sampvar, simweight = NULL,
  sampweight = NULL, realz = 1, ngrid = c(20, 20, 20),
  nquant = 1000, nvalues = 5000, logscale = FALSE, round = 2)
}
\arguments{
\item{sims}{fff}

\item{simvar}{fff}

\item{samples}{fff}

\item{sampvar}{fff}

\item{simweight}{fff}

\item{sampweight}{fff}

\item{realz}{fff}

\item{ngrid}{fff}

\item{nquant}{fff}

\item{nvalues}{fff}

\item{logscale}{fff}

\item{round}{fff}
}
\description{
\code{histpltsimR} is an interface to CCG Fortran program \code{histpltsim}.
It generates histogram data for multiple simulations and an optional
refernece dataset (e.g., samples). It is most useful for comparing the
simulated distribution with the input sample distribution as validation.
}
\details{
Functionality is different compared with the Fortran program. The following
differences are present:
\itemize{
\item Support for a file with lithology information.
\item Trimming limits are fixed.
\item Plot specific parameters title, stats position, and refernece value
for box plot are fixed.
\item Input data must be simulations, this is fixed.
\item Output always includes the numeric data, this is fixed.
\item Histograms always start at simulation 1 and end at \code{realz}.
\item Attribute minimum and maximum are derived form the data.}

These differences may be changed in future updates.
}
