% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usgsimR.R
\name{mvario_parstring}
\alias{mvario_parstring}
\title{Create GSLIB Parameter File Variogram Model String.}
\usage{
mvario_parstring(mvario)
}
\arguments{
\item{mvario}{A data frame of class variogramModel (\code{gstat} package).}
}
\value{
A character string suitable for use in GSLIB parameter files.
}
\description{
Create GSLIB Parameter File Variogram Model String.
}
