% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-write.R
\name{read_gslib}
\alias{read_gslib}
\title{Import GSLIB's simplified GeoEase data format to a data frame.}
\usage{
read_gslib(path)
}
\arguments{
\item{path}{Path or connection to read from.}
}
\value{
A data frame.
}
\description{
\code{read_gslib} imports from GeoEase format. All columns are read as
doubles as GSLIB does not support non-numeric data columns. The title is
appended to the data as a comment. Will import a GSLIB grid file but does
nothing with the grid definition information for now.
}
\examples{
read_gslib(rgslib_example("samples-na.dat"))
}
