% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varsimR.R
\name{read_gslib_grid_vario}
\alias{read_gslib_grid_vario}
\title{Import GSLIB-Format Gridded Experimental Data to a Data Frame.}
\usage{
read_gslib_grid_vario(path)
}
\arguments{
\item{path}{Path or connection to read from.}
}
\value{
A data frame.
}
\description{
\code{read_gslib_grid_vario} imports from GSLIB gridded experimental
variogram format producing a data frame of class 'gstatVariogram'.
}
