% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-write.R
\name{read_gslib_mvario}
\alias{read_gslib_mvario}
\title{Import GSLIB Variogram Model Format Data to a Data Frame.}
\usage{
read_gslib_mvario(path)
}
\arguments{
\item{path}{Path or connection to read from.}
}
\value{
A data frame.
}
\description{
\code{read_gslib_mvario} imports from GSLIB variogram model format. All
columns are read as doubles as GSLIB does not support non-numeric data
columns.
}
