% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-write.R
\name{read_gslib_usgsim}
\alias{read_gslib_usgsim}
\title{Import \code{usgsim} regular output format and add coordinates.}
\usage{
read_gslib_usgsim(path, vars)
}
\arguments{
\item{path}{Path or connection to read from.}

\item{vars}{Character vector of simulated column names.}
}
\value{
A data frame containing realization number, cartesian coordinates,
and simulated data.
}
\description{
\code{read_gslib_usgsim} imports the regular output format of
\code{usgsim}. This format has no grid coordinates. Coordinates are
added based on the grid definition on line 2 of the \code{usgsim} output.
}
