% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{rgslib_example}
\alias{rgslib_example}
\title{Get path to external example data.}
\usage{
rgslib_example(path = NULL)
}
\arguments{
\item{path}{Name of file. If \code{NULL}, the example files will be listed.}
}
\description{
\code{rgslib} comes bundled with some sample files in its \code{inst/extdata}
directory. This function make them easy to access.
}
\details{
This function is a direct modification of the function used by the \href{https://github.com/tidyverse/readr/blob/master/R/example.R}{readr package}
}
\examples{
rgslib_example()
rgslib_example("samples-na.dat")
}
