% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{sample_grid_extract}
\alias{sample_grid_extract}
\title{Extract Regulalarly Spaced Samples from a Grid}
\usage{
sample_grid_extract(grid, sample_def, xyz = c("x", "y"))
}
\arguments{
\item{grid}{Data fram input grid.}

\item{sample_def}{Standard grid definition, which is a named numeric vecotor
with elements (in order): n_x, n_y, n_z, min_x, min_y, min_z, dim_x,
dim_y, dim_z, realz. Used to define regular sample pattern.}

\item{xyz}{Coordinates in \code{grid}.}
}
\value{
Data frame of points sampled from \code{grid}.
}
\description{
Extract Regulalarly Spaced Samples from a Grid
}
