% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{samples_2d}
\alias{samples_2d}
\title{2D Point Sample Data.}
\format{A data frame with 44 rows and 6 variables:
\describe{
\item{id}{unique point ID}
\item{x}{x coordinate in metres}
\item{y}{y coordinate in metres}
\item{value}{value metal concentration in ppm}
\item{thk}{domain thickness in metres}
\item{accum}{metal accumulation in ppm·m}
}}
\usage{
data(samples_2d)
}
\description{
A dataset containing sample point data for a 2D (xy) domain. The samples
record a concentration of metal the domain thickness and the accumulation
(metal concentration × thickness).
}
\keyword{datasets}
