% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ublkavgR.R
\name{ublkavgR}
\alias{ublkavgR}
\title{Scale Grid Size Up or Down Using \code{ublkavg}.}
\usage{
ublkavgR(data, vars, domain = NULL, xyz = c("x", "y"),
  weights = NULL, factor = 2, minfrac = 1e-05, avg_method = 0,
  append_frac = FALSE, indims = c(1, 1, 1))
}
\arguments{
\item{data}{Data frame containing x, y, and, optionally z coordinates and one
or more fields to average or assign by dominance.}

\item{vars}{Character vecotr of variables to average or assign by dominance.}

\item{domain}{Character vector for domain field to assign by dominance.}

\item{xyz}{Character vector of coordinate field names.}

\item{weights}{Charcater vector of weighting field name(s).If one field it is
applied to all fields except \code{domain}.}

\item{factor}{Numeric positive number: multiplicative factor to scale grid.}

\item{minfrac}{Minimimum fraction of upscaled grid node that is represented
by the input grid.}

\item{avg_method}{Integer vector of averaging methods to use: 0 = arithmetic
mean; 1 = geometric mean; 2 = harmonic mean; 3 = majority rules (for
categorical variables)}

\item{append_frac}{Scalar boolean: append fraction field to output.}

\item{indims}{Numeric vector of grid dimensions for input grid.}
}
\value{
Data frame of re-scaled grid.
}
\description{
Scale Grid Size Up or Down Using \code{ublkavg}.
}
\details{
There is currently an issue with the implimentation of the \code{domain}
argument. It is supposed to produce a variable 0 to 1 in the output grid,
which is the average of a domain indicator. It current gives wrong result and
I am not sure if it is this implimentation of the Fortran program at fault.
}
