% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unscoreR.R
\name{unscoreR}
\alias{unscoreR}
\title{Run the CCG \code{unscore} Program to Normal Score Transform Sample Data.}
\usage{
unscoreR(data, vars, weight = NULL, cat = NULL, quantiles = 0,
  debug = FALSE)
}
\arguments{
\item{data}{Data frame with value and optional weight and category columns.}

\item{vars}{Character vector, column names in \code{data} for transformation.}

\item{weight}{Character scalar, name of weight field, default \code{NULL}.}

\item{cat}{Character scalar, name of category field, default \code{NULL}.}

\item{quantiles}{Integer scalar, number of quantiles output to transformation
tables, default 0, which means all.}

\item{debug}{Scalar boolean. If \code{TRUE} don't delete temporary system
files. Default is \code{FALSE}.}
}
\value{
A list of two data frames: \code{data} contains original input data
plus columns of normal score transformed data for each specified column and
\code{trans} contains the transformation tables for each transformed
variable.
}
\description{
\code{unscoreR} transforms selected variables to a normal distribution using
the CCG Fortran program \code{declus}. Currently the options to use a
category or to use a reference distribution are not supported. Trimming
limits are fixed to -1.0e21 and +1.0e21.
}
\examples{
samples_ns <- unscoreR(samples_2d, c("thk", "accum"))
par(mfrow=c(1,2))
hist(samples_ns$data$accum, breaks=30, main="raw", xlab="accum")
hist(samples_ns$data$NS_accum, main="normal", xlab="nscore accum")
}
