% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varcalcR.R
\name{varcalcR}
\alias{varcalcR}
\title{Run the CCG \code{varcalc} Program to Calculate Experimental Variograms.}
\usage{
varcalcR(data, vars, xyz = c("x", "y"), vpar = list(dir1 = c(azm = 0,
  azmtol = 22.5, bndhorz = 1000, dip = 0, diptol = 22.5, bndvert = 1000,
  tilt = 0, nlags = 10, dlag = 30, lagtol = 15), dir2 = c(azm = 90, azmtol
  = 22.5, bndhorz = 1000, dip = 0, diptol = 22.5, bndvert = 1000, tilt = 0,
  nlags = 10, dlag = 30, lagtol = 15), dir3 = c(azm = 0, azmtol = 22.5,
  bndhorz = 1000, dip = 90, diptol = 22.5, bndvert = 1000, tilt = 0, nlags
  = 10, dlag = 30, lagtol = 15)), standardize = 0, legacy = 0,
  check = 1, debug = FALSE)
}
\arguments{
\item{data}{Data frame of sample data.}

\item{vars}{Character vector, column names in \code{data} for variograms.}

\item{xyz}{Character vector, column names in \code{data} for coordinate
fields, 2 for 2D and 3 for 3D.}

\item{vpar}{list of named vectors containing parameters for each variogram
direction to be calculated. See function argument default for spec.}

\item{standardize}{Scalar integer either 0 or 1. 1: standardize the sill.}

\item{legacy}{Scalar integer either 0 or 1. 1: output legacy format.}

\item{check}{Scalar integer either 0 or 1. 1: run input checks.}

\item{debug}{Scalar boolean. If \code{TRUE} don't delete temporary system
files.}
}
\value{
A data frame with calculated variogram data.
}
\description{
\code{varcalcR} calculates experimental variograms. THis process is suited to
irregulalry spaced data such as samples. If more than one variable is input
all possible semi-variograms and cross-variograms are calculated. Trimming
limits are fixed to -1.0e21 and +1.0e21.
}
\examples{
test <- unscoreR(samples_2d, c("thk", "accum"))
data <- varcalcR(test$data, vars=c("NS_thk"))
}
