% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varmapR.R
\name{varmapR}
\alias{varmapR}
\title{Run the CCG \code{varmap} Program to Create a Variogram Map.}
\usage{
varmapR(data, vars, regular = 0, grid = c(0, 0, 0, 0, 0, 0),
  xyz = c("x", "y"), nlag = c(50, 50, 0), dlag = c(5, 5, 5),
  minpair = 5, standardize = 0, type = c(1, 1, 1), debug = FALSE)
}
\arguments{
\item{data}{Data frame with value and optional weight and category columns.}

\item{vars}{Character vector, column names in \code{data} for transformation.}

\item{regular}{Scalar integer either 1 or 0. 1: regular grid; 0: scattered
values.}

\item{grid}{Numeric vector of length 6. Defines the grid parameters if
\code{regular} is 1. The parameter are , in order, number of x, y, and z
nodes in the grid and size of x, y, and z nodes in the grid.}

\item{xyz}{Character vector of lenght 2 for 2D samples or 3 for 3D samples.
If \code{regular} is 0, \code{xyz} gives the x, y, and z coordinate
column names of \code{data}.}

\item{nlag}{Numeric vector of length 3 for number of lags in each direction.}

\item{dlag}{Numeric vector of length 3 for size of lag in each direction.}

\item{minpair}{Scalar numeric. Minimum pairs per lag to report a result.}

\item{standardize}{Scalar integer either 1 or 0. 1: standardize the sill.}

\item{type}{Numeric vector length 3 for the tail, head, and variogram type.}

\item{debug}{Scalar boolean. If \code{TRUE} don't delete temporary system
files.}
}
\value{
A data frame suitable for plotting with \code{ggplot2::geom_raster}.
}
\description{
\code{varmapR} creates variogram maps for analysis of spatial continuity.
The number of variogram maps produced is fixed at 1.
}
\examples{
test <- unscoreR(samples_2d, c("thk", "accum"))
data <- varmapR(test$data, vars=c("NS_thk"), minpair = 1,
  nlag = c(100, 100, 1), dlag = c(20, 20, 1))

}
