% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varmodelR.R
\name{varmodelR}
\alias{varmodelR}
\title{Run the CCG \code{varmodel} Program to Auto-Fit Variograms.}
\usage{
varmodelR(expdata, sysout, ndir = 1, nst = 10, c0 = "0.05:0.25",
  it = 1, cc = "?", azm = 0, dip = 0, tilt = 0, ahmax = "?",
  ahmin = "?", avert = "?", fit = TRUE, maxiter = 1e+05,
  vargsill = 1, vario = c(1, 2), npairswt = TRUE, invdistwt = TRUE,
  minpairs = 10, fixhmaxvertanis = FALSE, hmaxvertanis = 10,
  fixhmaxhminanis = FALSE, hmaxhminanis = 1)
}
\arguments{
\item{expdata}{Character: name of input GSLIB-format experiemental variogram
file without the extnsion part, and extension of ".out" is added for
compatibility with \code{vl_lmc}.}

\item{sysout}{Character: name of output GSLIB-format variogram model file
without the extnsion part, and extension of ".var" is added for
compatibility with \code{vl_lmc}.}

\item{ndir}{Numeric integer: number of directions to model.}

\item{nst}{Scalar integer: number of structures to model. Default 10.}

\item{c0}{Scalar numeric: nugget effect or wildcard.}

\item{it}{Scaler integer: structure type in range 1:5 or wildcard.}

\item{cc}{Scalar numeric: variance contribution or wildcard.}

\item{azm}{Numeric vector: azimuth in GSLIB convention or wildcard. Wildcard
not recommended. Must be length \code{ndir} with first value being major
axis azimuth.}

\item{dip}{Numeric vector: dip in GSLIB convention or wildcard. Wildcard
not recommended. Must be length \code{ndir} with first value being major
axis dip.}

\item{tilt}{Scalar numeric: tilt or major axis in GSLIB convention or
wildcard. Wildcard not recommended.}

\item{ahmax}{Scalar numeric: major axis range or wildcard.}

\item{ahmin}{Scalar numeric: semi-major axis range or wildcard.}

\item{avert}{Scalar numeric: minor axis range or wildcard.}

\item{fit}{Boolean: apply autofit.}

\item{maxiter}{Scalar integer: Number of fitting iterations.}

\item{vargsill}{Scalar numeric: sill to fit. Can be wildcard but not
recommended.}

\item{vario}{Numeric vector: variogram numbers in \code{expdata} to use.}

\item{npairswt}{Boolean: use number of pairs weighting.}

\item{invdistwt}{Boolean: use inverse distance weighting.}

\item{minpairs}{Scalar integer: minimum number of pairs.}

\item{fixhmaxvertanis}{Boolean: used fixed ahmax:avert anisotropy.}

\item{hmaxvertanis}{Scalar numeric: fixed ahmax:avert anisotropy.}

\item{fixhmaxhminanis}{Boolean: used fixed ahmax:ahmin anisotropy.}

\item{hmaxhminanis}{Scalar numeric: fixed ahmax:ahmin anisotropy.}
}
\value{
A data frame with discretized variogram models for display. The side
effect of this function is a system file {\code{sysout})} containing the
variogram model parameters for the major, semi-major, and minor axes. In
the case of a 2D variogram the semi-major and minor axes have the same
parameters.
}
\description{
\code{varmodelR} auto fits variogram models. This version is less flexible
the pure Fortran program. It is designed to support modelling of univariate
variogram models. For multivariate models it is designed to model the
direct variograms with the cross-variograms modeled in the \code{vl_lmc}
function. A large number of structures can be modelled (in additiona to the
nugget effect), but they all must have the same structure type, orientation,
and range parameters. If these parametrs are 'wildcards' they may produce
structures with different type and range.
}
\details{
Wildcards are character strings of the form '?': unconstrained; 'min:max':
constrained minimum and maximum; 'min:': constrained minimum; ':max':
constrained maximum. The arguments \code{c0}, \code{it}, \code{cc},
\code{azm}, \code{dip}, \code{tilt}, \code{ahmax}, \code{ahmin}, and
\code{avert} all support wildcards. The argument \code{c0} does not
support the '?' wildcard due to a possible bug in the Frotran program.
}
