% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varsimR.R
\name{varsimR}
\alias{varsimR}
\title{Only supports 2D rotations.
Does not support lithology data file.
Does not support indicator variograms.}
\usage{
varsimR(data, vars, xyz = c("x", "y"), dims = c(1, 1), realz = 1,
  standarize = FALSE, dirs = c(0, 90), nlags = 10, single = FALSE)
}
\arguments{
\item{data}{Data frame of gridded data.}

\item{vars}{Character vecotr of variables to calculate variograms.}

\item{xyz}{Character vecotr of coordinate column names.}

\item{dims}{Numeric vector of grid node/cell dimensions.}

\item{realz}{Scalar integer number of realizations.}

\item{standarize}{Scalar boolean: standardize variograms.}

\item{dirs}{Numeric vecotr of azimuth rotations.}

\item{nlags}{Number of lags to calculate.}

\item{single}{Boolean: produce a single variogram rather than direct and
cross variograms.}
}
\description{
Only supports 2D rotations.
Does not support lithology data file.
Does not support indicator variograms.
}
