% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vl_lmc.R
\name{vl_lmcR}
\alias{vl_lmcR}
\title{Run the CCG \code{vl_lmc} to Automatically Model Multivariate Variograms}
\usage{
vl_lmcR(nvars, corr0, corr, varcalcs = "vg_", varfits = "vm_",
  invdistwt = TRUE, npairswt = TRUE, minpairs = 10, maxiter = 100,
  varfitslmc = "vl_", corrs = "corrs.txt")
}
\arguments{
\item{nvars}{Scalar numeric: number of variables being modelled.}

\item{corr0}{Numeric vector: cross-variogram nugget effects.}

\item{corr}{Numeric matrix: correlation matrix (e.g., from \code{cor}).}

\item{varcalcs}{Character: prefix of experimental variogram files.}

\item{varfits}{Character: prefix of direct variogram model files.}

\item{invdistwt}{Boolean: inverse distance weighting.}

\item{npairswt}{Boolean: number of pairs weighting.}

\item{minpairs}{Scalar integer: minimum pairs.}

\item{maxiter}{Scalar integer: number of fitting iterations.}

\item{varfitslmc}{Character: prefix of output variogram model files.}

\item{corrs}{Character: output correlation matrix system file.}
}
\value{
A list with two data frames: \code{model} contains the variogram models
and \code{corrs} contains the model cross-variogram sills for validation.
}
\description{
\code{vl_lmcR} is an interface to the Fortran program \code{varfit_lmc}.
It auotmaticaly models direct and cross variograms given experimental points
and models fo the direc variograms.
}
