% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kt3dnR.R
\name{kt3dnR}
\alias{kt3dnR}
\title{Interface to \code{kt3dn} 3D Kriging Program.}
\usage{
kt3dnR(grid_def, samples, vars, bhid = NULL, xyz = c("x", "y", "z"),
  keygrid = NULL, keyval = 0, n_dataspac = 4, l_dataspac = 6,
  idbg = 0, ndis = c(3, 3, 1), minmax = c(16, 30), noct = 2,
  ndhmax = 0, srchdist = c(10, 10, 10), srchang = c(0, 0, 0),
  ktype = 1, skmean = 0, corr = 0, zvar = 0, mvario = NULL,
  idw = TRUE, debug = FALSE)
}
\arguments{
\item{grid_def}{Standard grid definition, which is a named numeric vector
with elements (in order): n_x, n_y, n_z, min_x, min_y, min_z, dim_x,
dim_y, dim_z, realz (realz is not used). This will be the grid that values
are estimated into.}

\item{samples}{Data frame of sample data.}

\item{vars}{Character vector (length 1--2) of variables in \code{samples} to
be estimated. First is primary, second is secondary.}

\item{bhid}{Scalar character name of hole id column in \code{samples}. Only
needed if \code{ndhmax} is > 0.}

\item{xyz}{Charcater vector of coordinate column names in \code{samples}.}

\item{keygrid}{A data frame grid with columns x, y, and z and a fourth
column containing a flag for nodes to be estimated (flag value defined in
argument \code{keyvalue}). This grid must have the same definition as that
defined in argument \code{grid_def}. The full extend of this grid
definition must be populated.}

\item{keyval}{Scalar integer key value in \code{keygrid} that indicates nodes
to be estiimated. Must be in column 4 after the three coordinate columns.}

\item{n_dataspac}{Scalar integer number fo drillhole neighbours for
estimation of 2D drillhole spacing.}

\item{l_dataspac}{Scalar numeric composite length to use when estimating
drillhole spacing in 3D data.}

\item{idbg}{Scalar integer debug level: 0 none; or 3, 5, 10.}

\item{ndis}{Integer vecots with number of discretization points in x, y and
z. Use all 1s for point kriging.}

\item{minmax}{Integer vector with minimum and maximum number of samples in
the search neighbourhood.}

\item{noct}{Scalar integer minimum samples for each octant. Use zero for no
octants.}

\item{ndhmax}{Scalar integer maximum samples per drillhole. Use zero for
deactivate. If > 0 must specify \code{bhid} column name. Only useful for 3D
kriging.}

\item{srchdist}{Numeric vecotor of search radii in rotated x, y, and z.}

\item{srchang}{Numeric vector of search oriention, azimuth, dip, and tilt.}

\item{ktype}{Scalar integer kriging type: 0=SK, 1=OK, 2=LVM(resid),
3=LVM((1-w)*m(u))) ,4=colo, 5=exdrift, 6=ICCK.}

\item{skmean}{Scalar numeric global mean grade for \code{ktype} 0, 4, 5, 6.}

\item{corr}{Scalar numeric correlation for \code{ktype} 4 or 6.}

\item{zvar}{Scalar numeric variance reduction factor for \code{ktype} 4.}

\item{mvario}{Variogram model data frame or list of models as \code{gstat}
class "variogramModel" or "variogramModelList".}

\item{idw}{Boolean scalar do inverse distance weighting instead of kriging.}

\item{debug}{Boolean scalar, if \code{TRUE} don't delete system files.}
}
\value{
Data frame grid with estimated variable.
}
\description{
Anomalies compared with raw program:
}
\details{
\itemize{
\item Cross and jacknife options not supported.
\item Auto search optimization not supported (will run if minimum set to
negative value, however no otion to change max and increment and only
first 'realization' is recognized by this implimentation.)
\item Kriging with external drift not supported  (will run if \code{ktype}
is set to 5, but no option to set drift indicators, etc.).
\item Input grid is required with a key value to define nodes to estimate.
}
}
