% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{structure_types}
\alias{structure_types}
\title{Make Data Frame Containing Valid GSLIB Variogram Model Structure Types}
\usage{
structure_types()
}
\value{
Data frame of structure types.
}
\description{
Links GSLIB numeric structure types to \code{gstat} short alphanumeric
structure types.
}
