% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usgsimR.R
\name{usgsimR}
\alias{usgsimR}
\title{Simulation Using \code{usgsim} Fortran Program from CCG.}
\usage{
usgsimR(data, mvario, vars, corr, xyz = c("x", "y"), n_realz = 1,
  seed = 60221409, xdef = c(50, 0.5, 1), ydef = c(50, 0.5, 1),
  zdef = c(50, 0.5, 1), simout = "sgsim", imputeout = "impute",
  debuglevel = 0, n_prev = 12, srchdist = c(10, 10, 10),
  srchang = c(0, 0, 0), sortmethod = 0, covarsort = c(1, 1, 1),
  clip = TRUE, assign = 1, trans = TRUE, nquant = 200,
  krigmethod = 2, altflag = 0, cosim = TRUE, domgrid = NULL)
}
\arguments{
\item{data}{Data frame containing coordinates and variables to condition the
simulations.}

\item{mvario}{Variogram model data frame or list of models as \code{gstat}
class "variogramModel" or "variogramModelList".}

\item{vars}{Character vector of column names to be simulated.}

\item{corr}{Correalation matrix (from \code{cor}) for \code{vars}.}

\item{xyz}{Character vector of coordinate column names.}

\item{n_realz}{Scalar integer number of simulation realizations.}

\item{seed}{Scalar integer seed for renadom number generation.}

\item{xdef}{Numeric vector with x-axis grid definition: number, origin, size.}

\item{ydef}{Numeric vector with y-axis grid definition: number, origin, size.}

\item{zdef}{Numeric vector with z-axis grid definition: number, origin, size.}

\item{simout}{Name of GeoEase grid file to contain output simulations.}

\item{imputeout}{Name of file to contain imputed data in heterotopic case.}

\item{debuglevel}{Scalar integer either 0 (none) or 1, 2, or 3 for increasing
levels of reporting.}

\item{n_prev}{Scalar integer number of previously simulated nodes to use.}

\item{srchdist}{Numeric vector search readii: x, y, z.}

\item{srchang}{Numeric vector search angles: ang1, ang2, ang3.}

\item{sortmethod}{Scalar numeric: sort by distance (0) or covariance (1).}

\item{covarsort}{Numeric vector: if sorting by covariance, indicate variogram
rock type, head, tail to use.}

\item{clip}{Scalar boolean: clip data to grid.}

\item{assign}{Scalar integer: assign to the grid, 0=none, 1=nearest, and
2=average.}

\item{trans}{Scalar boolean: perform normal score transormation of input
samples and backtransofrmation of simulation.}

\item{nquant}{Scalar integer: number of quantiles to keep from transform.}

\item{krigmethod}{Scalar integer: kriging method: 1=independent, 2=CCK, 4=CK,
5=ICCK, 6=BU.}

\item{altflag}{Scalar integer: option for primary variables if BU; or for
secondary variables if CK.}

\item{cosim}{Scalar boolean: perform cosimulation of multiple variables.}

\item{domgrid}{Data frame containing only x, y, and, optionally z coordinate
columns that define the domain grid. Output data will be restriceted to
these coorindate points. Coordinate columns must be the same as \code{xyz}.}
}
\value{
Data frame of simulation results.
}
\description{
\code{usgsimR} is an interface to the CCG simulation program \code{usgsim},
which uses SGS. Some functionality of the Fortran program is not implimented
in this interface:
- Weight field.
- Rock types.
- Seconday variables.
- Normal-score transformation reference distributions.
- Trend data.
- Only writes output format 0: regular.
}
