% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varcalcR.R
\name{varcalcR}
\alias{varcalcR}
\title{Run the CCG \code{varcalc} Program to Calculate Experimental Variograms.}
\usage{
varcalcR(data, vars, sysout, xyz = c("x", "y"), vpar = list(dir1 =
  c(azm = 0, azmtol = 90, bandhorz = 1e+21, dip = 0, diptol = 90, bandvert
  = 1e+21, tilt = 0, nlags = 10, lagdist = 30, lagtol = 0.5)),
  variostd = FALSE, variosills = NULL, legacy = FALSE,
  strict = TRUE, single = TRUE, debug = TRUE)
}
\arguments{
\item{data}{Data frame of sample data.}

\item{vars}{Character vector, column names in \code{data} for variograms.}

\item{sysout}{Character: name of GSLIB-format experimental variogram file
without the extension part, and extension of ".out" is added for
compatibility with \code{vl_lmc}.}

\item{xyz}{Character vector, column names in \code{data} for coordinate
fields, 2 for 2D and 3 for 3D.}

\item{vpar}{list of named vectors containing parameters for each variogram
direction to be calculated. See function argument default for spec.}

\item{variostd}{Scalar boolean, if \code{TRUE} standardize the sill.}

\item{variosills}{Numeric scalar. Sill to standardize to if \code{variostd}
is \code{TRUE}. Default \code{NULL} means let program decide.}

\item{legacy}{Scalar boolean, if \code{TRUE} output legacy format.}

\item{strict}{Scalar boolean, if \code{TRUE} run input checks.}

\item{single}{Boolean: produce a single variogram rather than direct and
cross variograms.}

\item{debug}{Scalar boolean. If \code{TRUE} don't delete temporary system
files.}
}
\value{
A data frame of class "gstat::gstatVariogram" with calculated
variogram data. Side effect is output variogram calculation file from
\code{varcalc} program. This file is used by other programs such as
\code{varmod}.
}
\description{
\code{varcalcR} calculates experimental variograms. This process is suited to
irregulalry spaced data such as samples. If more than one variable is input
all possible semi-variograms and cross-variograms are calculated. Default
arguments will create omni-directional 2D variograms for each variable.
}
\details{
Variogram calculation parameters are provided as a list of vectors for each
direction. The default list provided in the function argument list uses a
named vector. The vector does not need ot me named but values must be
provided in the order shown.
}
