% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-write.R
\name{write_gslib}
\alias{write_gslib}
\title{Export a Data Frame to Simplified GeoEase Format Used by GSLIB.}
\usage{
write_gslib(data, path, round = 4, title = NULL)
}
\arguments{
\item{data}{Data frame to write to disk, only numeric columns are exported.}

\item{path}{Path or connection to write to.}

\item{round}{Integer scalar, number of digits to round data before export.}

\item{title}{Character title to be appended to GeoEase file. If not supplied
it will be determined from \code{data} meta-data or it's name.}
}
\value{
\code{write_gslib} returns the input \code{data} invisibly.
}
\description{
\code{write_gslib} writes a data frame to the GeoEase format used by GSLIB
programes. GeoEase has a standard header with a title and number and names of
the columns. The title can be supplied or is built automatically from meta-
data in the data frame or the name of the data fram object. Only numeric
columns are written to the system file and they are rounded. This version
not support wrting of the grid-format GSLIB data file.
}
\examples{
samples_na <- samples_2d
samples_na[3:7, 5] <- NA
write_gslib(samples_na, "samples-na.dat", round = 3, title = "sample data")
}
